RF: move all the activation functionality into separate module
